feat: set final ID for VIP-199, and description
